+++ lib/exceptions/errors.dart
