+++ lib/protocols/text_message_format.dart
