+++ lib/signalr_client.dart
@@ -1,9 +1,9 @@
 library signalr_client;
 
-export 'web_supporting_http_client.dart';
-export 'http_connection.dart';
-export 'http_connection_options.dart';
-export 'hub_connection.dart';
-export 'hub_connection_builder.dart';
-export 'itransport.dart';
-export 'signalr_http_client.dart';
+export 'clients/signalr_http_client.dart';
+export 'clients/web_supporting_http_client.dart';
+export 'connection/http_connection.dart';
+export 'connection/http_connection_options.dart';
+export 'connection/hub_connection.dart';
+export 'connection/hub_connection_builder.dart';
+export 'transport/itransport.dart';
